Refactor: 특일정보 api 리팩토링
