+++ core/data/src/main/java/kr/co/data/mapper/HolidayMapper.kt
@@ -1,16 +1,16 @@
 package kr.co.data.mapper
 
-import kr.co.common.mapper.Mapper
-import kr.co.data.model.data.HolidayData
+import kr.co.common.mapper.BaseMapper
+import kr.co.data.model.data.calendar.HolidayData
 import kr.co.domain.entity.HolidayEntity
 import kr.co.domain.entity.type.HolidayType
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 
 internal object HolidayMapper :
-    Mapper<HolidayData, HolidayEntity> {
-    override fun convert(param: HolidayData): HolidayEntity =
-        with(param) {
+    BaseMapper<HolidayData, HolidayEntity>() {
+    override fun toRight(left: HolidayData): HolidayEntity =
+        with(left) {
             HolidayEntity(
                 date = LocalDate.parse(date, DateTimeFormatter.ofPattern("yyyyMMdd")),
                 isHoliday = when (isHoliday) {
@@ -20,6 +20,16 @@ internal object HolidayMapper :
                 },
                 type = HolidayType.ofValue(type),
                 name = name,
+            )
+        }
+
+    override fun toLeft(right: HolidayEntity): HolidayData =
+        with(right){
+            HolidayData(
+                date = date.format(DateTimeFormatter.ofPattern("yyyyMMdd")),
+                isHoliday = if (isHoliday) "Y" else "N",
+                type = type.typeString,
+                name = name
             )
         }
 }
