+++ core/data/src/main/java/kr/co/data/model/data/calendar/HolidayData.kt
@@ -1,4 +1,4 @@
-package kr.co.data.model.data
+package kr.co.data.model.data.calendar
 
 data class HolidayData(
     val date: String,
