+++ core/data/src/main/java/kr/co/data/source/remote/HolidayRemoteDataSource.kt
@@ -1,7 +1,7 @@
 package kr.co.data.source.remote
 
 import kotlinx.coroutines.flow.Flow
-import kr.co.data.model.data.HolidayData
+import kr.co.data.model.data.calendar.HolidayData
 
 interface HolidayRemoteDataSource {
     suspend fun fetchList(year: String, month: String): Flow<List<HolidayData>>
